+++ src/main_window.rs
@@ -6,15 +6,18 @@ extern crate nfd;
 extern crate sixtyfps;
 
 use sixtyfps::{Model, ModelHandle, SharedString};
-use std::collections::HashMap;
 use std::fmt::Debug;
 use std::path::Path;
 use std::rc::Rc;
+use std::sync::Arc;
 use std::sync::Mutex;
 use std::thread;
-use std::{cell::RefCell, sync::Arc};
 
+use crate::gui_items::list_item_from_file_item;
+use crate::gui_items::list_item_title;
+use crate::gui_items::sort_item_from_file_item;
 use crate::item_sort_list::parse_date;
+use crate::item_sort_list::FileItem;
 use crate::item_sort_list::ItemList;
 use crate::misc::image_cache::{self, ImageCache, Purpose};
 use crate::persistence::json::JsonPersistence;
@@ -36,15 +39,12 @@ mod generated_code {
 }
 pub use generated_code::*;
 
-type ImagesModelMap = HashMap<usize, usize>;
-
 /// Main window container of the image sorter, contains the sixtyfps window, models and internal data structures
 pub struct MainWindow {
     window: ImageSieve,
     item_list: Arc<Mutex<ItemList>>,
-    item_list_model: Rc<sixtyfps::VecModel<SharedString>>,
-    similar_items_model: Rc<sixtyfps::VecModel<SortImage>>,
-    items_model_map: Rc<RefCell<ImagesModelMap>>,
+    list_model: Rc<sixtyfps::VecModel<ListItem>>,
+    similar_images_model: Rc<sixtyfps::VecModel<SortItem>>,
     events_model: Rc<sixtyfps::VecModel<Event>>,
     sieve_result_model: Rc<sixtyfps::VecModel<SieveResult>>,
     image_cache: Rc<ImageCache>,
@@ -78,7 +78,7 @@ impl MainWindow {
         let item_list = Arc::new(Mutex::new(item_list));
 
         let event_list_model = Rc::new(sixtyfps::VecModel::<Event>::default());
-        let item_list_model = Rc::new(sixtyfps::VecModel::<SharedString>::default());
+        let item_list_model = Rc::new(sixtyfps::VecModel::<ListItem>::default());
         let sieve_result_model = Rc::new(sixtyfps::VecModel::<SieveResult>::default());
 
         // Construct main window
@@ -98,9 +98,8 @@ impl MainWindow {
         let main_window = Self {
             window: image_sieve,
             item_list,
-            item_list_model,
-            similar_items_model: Rc::new(sixtyfps::VecModel::<SortImage>::default()),
-            items_model_map: Rc::new(RefCell::new(HashMap::new())),
+            list_model: item_list_model,
+            similar_images_model: Rc::new(sixtyfps::VecModel::<SortItem>::default()),
             events_model: event_list_model,
             sieve_result_model,
             image_cache: Rc::new(cache),
@@ -121,13 +120,11 @@ impl MainWindow {
         // Set model references
         main_window
             .window
-            .set_images_list_model(sixtyfps::ModelHandle::new(
-                main_window.item_list_model.clone(),
-            ));
+            .set_list_model(sixtyfps::ModelHandle::new(main_window.list_model.clone()));
         main_window
             .window
-            .set_images_model(sixtyfps::ModelHandle::new(
-                main_window.similar_items_model.clone(),
+            .set_similar_images_model(sixtyfps::ModelHandle::new(
+                main_window.similar_images_model.clone(),
             ));
         main_window
             .window
@@ -165,20 +162,22 @@ impl MainWindow {
         self.window.on_item_selected({
             // New item selected on the list of images or next/previous clicked
             let item_list = self.item_list.clone();
-            let similar_items_model = self.similar_items_model.clone();
-            let items_model_map = self.items_model_map.clone();
+            let list_model = self.list_model.clone();
+            let similar_items_model = self.similar_images_model.clone();
             let window_weak = self.window.as_weak();
             let image_cache = self.image_cache.clone();
 
             move |i: i32| {
-                synchronize_images_model(
-                    i as usize,
-                    &item_list.lock().unwrap(),
+                let index = list_model.row_data(i as usize).local_index as usize;
+                let item_list = &item_list.lock().unwrap();
+                synchronize_similar_items_model(
+                    index,
+                    item_list,
                     similar_items_model.clone(),
-                    &mut items_model_map.borrow_mut(),
                     window_weak.clone(),
                     &image_cache,
                 );
+                prefetch_images(i, item_list, &list_model, &image_cache);
             }
         });
 
@@ -197,34 +196,38 @@ impl MainWindow {
             }
         });
 
-        self.window.on_take_over_toggle({
+        self.window.on_set_take_over({
             // Image was clicked, toggle take over state
-            let item_list_model = self.item_list_model.clone();
+            let list_model = self.list_model.clone();
+            let similar_items_model = self.similar_images_model.clone();
             let item_list = self.item_list.clone();
-            let items_model = self.similar_items_model.clone();
-            let items_model_map = self.items_model_map.clone();
 
-            move |i: i32| {
-                let model_index = i as usize;
+            move |i: i32, take_over: bool| {
                 // Change the state of the SortImage in the items_model
-                let mut sort_image = items_model.row_data(model_index);
-                sort_image.take_over = !sort_image.take_over;
 
-                let index = items_model_map.borrow_mut()[&model_index];
+                let index = i as usize;
                 {
                     // Change the item_list state
                     let mut item_list_mut = item_list.lock().unwrap();
-                    item_list_mut.items[index].set_take_over(sort_image.take_over);
+                    item_list_mut.items[index].set_take_over(take_over);
                 }
-                items_model.set_row_data(model_index, sort_image);
                 // Update item list model to reflect change in icons in list
-                synchronize_item_list_model(&item_list.lock().unwrap(), &item_list_model);
+                update_list_model_texts(&item_list.lock().unwrap(), &list_model);
+                // And update the take over state in the similar items model
+                for count in 0..similar_items_model.row_count() {
+                    let mut item: SortItem = similar_items_model.row_data(count);
+                    if item.local_index == i {
+                        item.take_over = take_over;
+                        similar_items_model.set_row_data(count, item);
+                        break;
+                    }
+                }
             }
         });
 
         self.window.on_browse_source({
             // Browse source was clicked, select new path
-            let item_list_model = self.item_list_model.clone();
+            let list_model = self.list_model.clone();
             let events_model = self.events_model.clone();
             let item_list = self.item_list.clone();
             let window_weak = self.window.as_weak();
@@ -249,7 +252,7 @@ impl MainWindow {
                         }
                     }
 
-                    empty_model(item_list_model.clone());
+                    empty_model(list_model.clone());
                     empty_model(events_model.clone());
 
                     // Synchronize in a background thread
@@ -317,7 +320,7 @@ impl MainWindow {
 
         self.window.on_add_event({
             // New event was added, return true if the dates are ok
-            let item_list_model = self.item_list_model.clone();
+            let list_model = self.list_model.clone();
             let events_model = self.events_model.clone();
             let item_list = self.item_list.clone();
 
@@ -332,9 +335,9 @@ impl MainWindow {
                 let mut item_list = item_list.lock().unwrap();
                 item_list.events.push(event);
                 item_list.events.sort_unstable();
-                synchronize_event_list_model(&item_list, &events_model);
+                synchronize_event_model(&item_list, &events_model);
                 // Synchronize the item list to update the icons of the entries
-                synchronize_item_list_model(&item_list, &item_list_model.clone());
+                update_list_model_texts(&item_list, &list_model.clone());
             }
         });
 
@@ -343,7 +346,7 @@ impl MainWindow {
         });
 
         self.window.on_update_event({
-            let item_list_model = self.item_list_model.clone();
+            let list_model = self.list_model.clone();
             let events_model = self.events_model.clone();
             let item_list = self.item_list.clone();
 
@@ -357,15 +360,15 @@ impl MainWindow {
                     event.end_date.as_str(),
                 ) {
                     item_list.events.sort_unstable();
-                    synchronize_event_list_model(&item_list, &events_model);
-                    synchronize_item_list_model(&item_list, &item_list_model.clone());
+                    synchronize_event_model(&item_list, &events_model);
+                    update_list_model_texts(&item_list, &list_model.clone());
                 }
             }
         });
 
         self.window.on_remove_event({
             // Event was removed
-            let item_list_model = self.item_list_model.clone();
+            let list_model = self.list_model.clone();
             let events_model = self.events_model.clone();
             let item_list = self.item_list.clone();
 
@@ -374,7 +377,7 @@ impl MainWindow {
                 let mut item_list = item_list.lock().unwrap();
                 item_list.events.remove(index as usize);
                 // Synchronize the item list to update the icons of the entries
-                synchronize_item_list_model(&item_list, &item_list_model.clone());
+                update_list_model_texts(&item_list, &list_model.clone());
             }
         });
 
@@ -411,40 +414,68 @@ impl MainWindow {
                 synchronizer.stop();
             }
         });
+
+        self.window.on_filter({
+            let list_model = self.list_model.clone();
+            let item_list = self.item_list.clone();
+            let window_weak = self.window.as_weak();
+
+            move |filters| {
+                let item_list = item_list.lock().unwrap();
+                empty_model(list_model.clone());
+                populate_list_model(&item_list, &list_model, &filters);
+                let rows = list_model.row_count() as i32;
+
+                if rows >= window_weak.unwrap().get_current_list_item() {
+                    window_weak.unwrap().set_current_list_item(rows - 1);
+                }
+            }
+        });
     }
 }
 
-fn empty_model<T: 'static + Clone>(item_list_model: Rc<sixtyfps::VecModel<T>>) {
-    for _ in 0..item_list_model.row_count() {
-        item_list_model.remove(0);
+/// Removes all items from a model
+fn empty_model<T: 'static + Clone>(vec_model: Rc<sixtyfps::VecModel<T>>) {
+    for _ in 0..vec_model.row_count() {
+        vec_model.remove(0);
     }
 }
 
-/// Synchronizes the list of found items from the internal data structure with the sixtyfps VecModel
-pub fn synchronize_item_list_model(
+/// Fills the list of found items from the internal data structure to the sixtyfps VecModel
+pub fn populate_list_model(
     item_list: &ItemList,
-    item_list_model: &sixtyfps::VecModel<SharedString>,
+    list_model: &sixtyfps::VecModel<ListItem>,
+    filters: &Filters,
 ) {
-    let empty_model = item_list_model.row_count() == 0;
-    for (index, image) in item_list.items.iter().enumerate() {
-        let mut item_string = image.get_item_string(&item_list.path);
-        if item_list.get_event(image).is_some() {
-            item_string = String::from("\u{1F4C5}") + &item_string;
-        }
-        if empty_model {
-            item_list_model.push(SharedString::from(item_string));
-        } else {
-            item_list_model.set_row_data(index, SharedString::from(item_string));
-        }
+    let mut filtered_list: Vec<&FileItem> = item_list
+        .items
+        .iter()
+        .filter(|item| filter_file_items(item, filters))
+        .collect();
+    filtered_list.sort_unstable_by(|a, b| compare_file_items(a, b, filters));
+    if filters.direction == "Desc" {
+        filtered_list.reverse();
+    }
+    for image in filtered_list {
+        let list_item = list_item_from_file_item(image, item_list);
+        list_model.push(list_item);
+    }
+}
+
+/// Update the texts for all entries in the list model
+/// Should be called when the underlying data (i.e. the item list) has changed
+pub fn update_list_model_texts(item_list: &ItemList, list_model: &sixtyfps::VecModel<ListItem>) {
+    for count in 0..list_model.row_count() {
+        let mut list_item = list_model.row_data(count);
+        let file_item = &item_list.items[list_item.local_index as usize];
+        list_item.text = list_item_title(file_item, item_list);
+        list_model.set_row_data(count, list_item);
     }
 }
 
 /// Synchronize the event list with the GUI model
-pub fn synchronize_event_list_model(
-    item_list: &ItemList,
-    event_list_model: &sixtyfps::VecModel<Event>,
-) {
-    let model_count = event_list_model.row_count();
+pub fn synchronize_event_model(item_list: &ItemList, event_model: &sixtyfps::VecModel<Event>) {
+    let model_count = event_model.row_count();
     // Event model
     for (index, event) in item_list.events.iter().enumerate() {
         let _event = Event {
@@ -453,34 +484,31 @@ pub fn synchronize_event_list_model(
             end_date: SharedString::from(event.end_date_as_string()),
         };
         if index >= model_count {
-            event_list_model.push(_event);
+            event_model.push(_event);
         } else {
-            event_list_model.set_row_data(index, _event);
+            event_model.set_row_data(index, _event);
         }
     }
 }
 
 /// Synchronizes the images to show at the same time from a selected image to the sixtyfps VecModel
-fn synchronize_images_model(
-    selected_item_index: usize,
+fn synchronize_similar_items_model(
+    current_item_local_index: usize,
     item_list: &ItemList,
-    similar_items_model: Rc<sixtyfps::VecModel<SortImage>>,
-    item_model_map: &mut ImagesModelMap,
+    similar_items_model: Rc<sixtyfps::VecModel<SortItem>>,
     window: sixtyfps::Weak<ImageSieve>,
     image_cache: &ImageCache,
 ) {
-    let similars = item_list.items[selected_item_index].get_similars();
+    let similars = item_list.items[current_item_local_index].get_similars();
 
-    // Clear images model and the model map
-    for _ in 0..similar_items_model.row_count() {
-        similar_items_model.remove(0);
-    }
-    item_model_map.drain();
+    // Clear images model
+    empty_model(similar_items_model.clone());
 
     let mut model_index: usize = 0;
 
     let mut add_item = |item_index: &usize,
-                        selected_image: bool,
+                        current_image: bool,
+                        has_similars: bool,
                         window_weak: sixtyfps::Weak<ImageSieve>| {
         let item = &item_list.items[*item_index];
         let image = {
@@ -490,14 +518,20 @@ fn synchronize_images_model(
             } else {
                 let f: image_cache::DoneCallback = Box::new(move |image_buffer| {
                     window_weak.clone().upgrade_in_event_loop(move |handle| {
-                        if handle.get_current_list_item() == selected_item_index as i32 {
-                            let mut row_data = handle.get_images_model().row_data(model_index);
-                            let is_current_image = handle.get_current_image().text == row_data.text;
-                            row_data.image = crate::misc::images::get_sixtyfps_image(&image_buffer);
-                            handle
-                                .get_images_model()
-                                .set_row_data(model_index, row_data);
-                            if is_current_image {
+                        // Check if still the image is visible that caused the image loads
+                        if handle.get_current_image().local_index == current_item_local_index as i32
+                        {
+                            let mut row_data =
+                                handle.get_similar_images_model().row_data(model_index);
+                            if has_similars {
+                                row_data.image =
+                                    crate::misc::images::get_sixtyfps_image(&image_buffer);
+                                handle
+                                    .get_similar_images_model()
+                                    .set_row_data(model_index, row_data);
+                            }
+                            // If the image is the current image, then we need to also update the current image SortImage
+                            if current_image {
                                 let mut current_image = handle.get_current_image();
                                 current_image.image =
                                     crate::misc::images::get_sixtyfps_image(&image_buffer);
@@ -508,8 +542,8 @@ fn synchronize_images_model(
                 });
                 image_cache.load(
                     item,
-                    if selected_image {
-                        Purpose::SelectedImage
+                    if current_image {
+                        Purpose::CurrentImage
                     } else {
                         Purpose::SimilarImage
                     },
@@ -519,55 +553,47 @@ fn synchronize_images_model(
             }
         };
 
-        let sort_image_struct = SortImage {
-            image,
-            take_over: item.get_take_over(),
-            text: get_item_text(*item_index, item_list),
-        };
-        similar_items_model.push(sort_image_struct);
-        item_model_map.insert(model_index, *item_index);
+        let sort_image = sort_item_from_file_item(item, item_list, image);
+        similar_items_model.push(sort_image);
         model_index += 1;
     };
 
-    // TODO: Also first item should be loaded in background, but in a prioritized queue
-    add_item(&selected_item_index, true, window.clone());
+    // Clear pending commands in the image cache
+    image_cache.purge();
 
+    add_item(
+        &current_item_local_index,
+        true,
+        !similars.is_empty(),
+        window.clone(),
+    );
     for image_index in similars {
-        add_item(image_index, false, window.clone());
-    }
-
-    // Prefetch next two images
-    let mut prefetch_index = selected_item_index + 1;
-    let mut prefetches = 2;
-    while prefetches > 0 && prefetch_index < item_list.items.len() {
-        if !similars.contains(&prefetch_index) {
-            if let Some(file_item) = item_list.items.get(prefetch_index) {
-                if file_item.is_image() {
-                    image_cache.load(file_item, Purpose::Prefetch, None);
-                    prefetches -= 1;
-                }
-            }
-        }
-        prefetch_index += 1;
+        add_item(image_index, false, !similars.is_empty(), window.clone());
     }
 
     // Set properties
     window
         .unwrap()
         .set_current_image(similar_items_model.row_data(0));
-    window.unwrap().set_current_image_index(0);
 }
 
-/// Gets the text of a an item at a given index as a SharedString
-pub fn get_item_text(index: usize, item_list: &ItemList) -> SharedString {
-    let item = &item_list.items[index];
-    let event = item_list.get_event(item);
-    let event_str = if let Some(event) = event {
-        event.name.as_str()
-    } else {
-        ""
-    };
-    SharedString::from(item.to_string() + " " + event_str)
+/// Prefetch the next images in the model list
+fn prefetch_images(
+    model_index: i32,
+    item_list: &ItemList,
+    list_model: &sixtyfps::VecModel<ListItem>,
+    image_cache: &ImageCache,
+) {
+    // Prefetch next two images
+    for i in model_index + 1..model_index + 3 {
+        if i < list_model.row_count() as i32 {
+            let list_item = &list_model.row_data(i as usize);
+            let file_item = &item_list.items[list_item.local_index as usize];
+            if file_item.is_image() {
+                image_cache.load(file_item, Purpose::Prefetch, None);
+            }
+        }
+    }
 }
 
 /// Sieves the item list in a background thread
@@ -635,4 +661,38 @@ pub fn sieve(
             progress_callback,
         );
     });
+}
+
+/// Filter file items to display in the item list
+fn filter_file_items(file_item: &FileItem, filters: &Filters) -> bool {
+    let mut visible = true;
+    if !filters.images && file_item.is_image() {
+        visible = false;
+    }
+    if !filters.videos && file_item.is_video() {
+        visible = false;
+    }
+    if !filters.sorted_out && !file_item.get_take_over() {
+        visible = false;
+    }
+    visible
+}
+
+/// Compare two file items taking the current sort settings into account
+fn compare_file_items(a: &FileItem, b: &FileItem, filters: &Filters) -> std::cmp::Ordering {
+    match filters.sort_by.as_str() {
+        "Date" => a.cmp(b),
+        "Name" => a.path.cmp(&b.path),
+        "Type" => {
+            if a.is_image() && b.is_image() {
+                a.cmp(b)
+            } else if a.is_image() && b.is_video() {
+                std::cmp::Ordering::Less
+            } else {
+                std::cmp::Ordering::Greater
+            }
+        }
+        "Size" => a.get_size().cmp(&b.get_size()),
+        _ => std::cmp::Ordering::Less,
+    }
 }