+++ src/synchronize.rs
@@ -1,4 +1,5 @@
 use crate::item_sort_list::ItemList;
+use crate::main_window::update_list_model_texts;
 use crate::persistence::settings::Settings;
 use image::GenericImageView;
 use img_hash::HashAlg;
@@ -12,7 +13,7 @@ use sixtyfps::VecModel;
 use walkdir::WalkDir;
 
 use crate::main_window::{
-    synchronize_event_list_model, synchronize_item_list_model, Event, ImageSieve, SortImage,
+    populate_list_model, synchronize_event_model, Event, ImageSieve, ListItem, SortItem,
 };
 use crate::misc::images::get_empty_image;
 use crate::persistence::json::get_project_filename;
@@ -124,11 +125,11 @@ fn synchronize_run(
         image_sieve.clone().upgrade_in_event_loop({
             let item_list = item_list.lock().unwrap().to_owned();
             move |h| {
-                synchronize_item_list_model(
+                update_list_model_texts(
                     &item_list,
-                    h.get_images_list_model()
+                    h.get_list_model()
                         .as_any()
-                        .downcast_ref::<VecModel<SharedString>>()
+                        .downcast_ref::<VecModel<ListItem>>()
                         .unwrap(),
                 );
                 h.set_calculating_similarities(false);
@@ -184,14 +185,16 @@ fn update_item_list(item_list: Arc<Mutex<ItemList>>, image_sieve: &sixtyfps::Wea
         let item_list = item_list.lock().unwrap().to_owned();
         let flag = flag.clone();
         move |h| {
-            synchronize_item_list_model(
+            let filters = h.get_filters();
+            populate_list_model(
                 &item_list,
-                h.get_images_list_model()
+                h.get_list_model()
                     .as_any()
-                    .downcast_ref::<VecModel<SharedString>>()
+                    .downcast_ref::<VecModel<ListItem>>()
                     .unwrap(),
+                &filters,
             );
-            synchronize_event_list_model(
+            synchronize_event_model(
                 &item_list,
                 h.get_events_model()
                     .as_any()
@@ -205,17 +208,17 @@ fn update_item_list(item_list: Arc<Mutex<ItemList>>, image_sieve: &sixtyfps::Wea
                 h.set_current_list_item(0);
                 h.invoke_item_selected(0);
             } else {
-                let empty_image = SortImage {
+                let empty_image = SortItem {
                     image: get_empty_image(),
                     take_over: true,
                     text: SharedString::from("No images found"),
+                    local_index: 0,
                 };
                 h.set_current_image(empty_image);
-                h.set_current_image_index(0);
-                let model_handle = h.get_images_model();
+                let model_handle = h.get_similar_images_model();
                 let images_list_model = model_handle
                     .as_any()
-                    .downcast_ref::<VecModel<SortImage>>()
+                    .downcast_ref::<VecModel<SortItem>>()
                     .unwrap();
                 for _ in 0..images_list_model.row_count() {
                     images_list_model.remove(0);
@@ -261,11 +264,11 @@ fn calculate_similar_timestamps(
         let item_list = item_list.lock().unwrap().to_owned();
         let use_hash = settings.use_hash;
         move |h| {
-            synchronize_item_list_model(
+            update_list_model_texts(
                 &item_list,
-                h.get_images_list_model()
+                h.get_list_model()
                     .as_any()
-                    .downcast_ref::<VecModel<SharedString>>()
+                    .downcast_ref::<VecModel<ListItem>>()
                     .unwrap(),
             );
             h.set_calculating_similarities(use_hash);