+++ CHANGELOG.md
@@ -6,10 +6,11 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
  - Now reading EXIF files for all image formats, not only for JPEGs
+ - Optimized loading of similar images
 
 ### Added
 
- - 
+ - Added filters and sort options to the item list
 
 ### Fixed